engine/requirements_gpu.txt
pandas
SQLAlchemy
psycopg2-binary
kagglehub
python-dotenv
tqdm
scikit-learn
xgboost
transformers
spacy
https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl
torch
nltk